Refactor: separate Location component + useLocation hook
