+++ QuickStartTemplate/projects/QuickStartTemplate-frontend/src/components/sunset/HeroSection.tsx
@@ -1,5 +1,7 @@
 import React, { useEffect, useRef, useState } from 'react'
 import MapFullScreen from './MapFullScreen'
+import LocationSelector from './LocationSelector'
+import { useLocation } from '../../hooks/useLocation'
 
 const VITE_OPENAI_KEY = import.meta.env.VITE_OPENAI_API_KEY || ''
 const VITE_OPENAI_MODEL = import.meta.env.VITE_OPENAI_MODEL || 'gpt-4o-mini'
@@ -7,118 +9,35 @@ const VITE_OPENAI_MODEL = import.meta.env.VITE_OPENAI_MODEL || 'gpt-4o-mini'
 interface HeroSectionProps {}
 
 const HeroSection: React.FC<HeroSectionProps> = () => {
-  const [location, setLocation] = useState<string>('')
-  const [isLoadingLocation, setIsLoadingLocation] = useState<boolean>(false)
-  const [showLocationDropdown, setShowLocationDropdown] = useState<boolean>(false)
-  const [customLocation, setCustomLocation] = useState<string>('')
-  const [hasLocationBeenSet, setHasLocationBeenSet] = useState<boolean>(false)
+  // Location logic extracted to custom hook
+  const {
+    location,
+    coords,
+    isLoadingLocation,
+    showLocationDropdown,
+    customLocation,
+    hasLocationBeenSet,
+    setLocation,
+    setShowLocationDropdown,
+    setCustomLocation,
+    getUserLocation,
+    handleLocationSelect,
+    handleCustomLocationSubmit,
+    geocodeLocation,
+  } = useLocation()
+
+  // Other state variables remain the same
   const [openMap, setOpenMap] = useState<boolean>(false)
   const [sunsetProbability, setSunsetProbability] = useState<number | null>(null)
   const [sunsetDescription, setSunsetDescription] = useState<string>('')
   const [sunsetLoading, setSunsetLoading] = useState(false)
   const [sunsetError, setSunsetError] = useState<string | null>(null)
-  const [coords, setCoords] = useState<{ lat: number; lon: number } | null>(null)
   const [weatherSummary, setWeatherSummary] = useState<string>('') // optional debug / context
 
   const debounceRef = useRef<ReturnType<typeof setTimeout> | null>(null)
   const lastSaltRef = useRef<string>('')
   const previousResultRef = useRef<{ location: string; probability: number | null } | null>(null)
 
-  // Popular locations for dropdown
-  const popularLocations = [
-    'Berlin, Germany',
-    'New York, USA',
-    'Tokyo, Japan',
-    'London, UK',
-    'Paris, France',
-    'Sydney, Australia',
-    'Toronto, Canada',
-    'Dubai, UAE',
-    'Singapore',
-    'Mumbai, India',
-  ]
-
-  // Get user's location using Geolocation API
-  const getUserLocation = () => {
-    setIsLoadingLocation(true)
-
-    if ('geolocation' in navigator) {
-      navigator.geolocation.getCurrentPosition(
-        async (position) => {
-          const { latitude, longitude } = position.coords
-          setCoords({ lat: latitude, lon: longitude })
-
-          try {
-            // Use reverse geocoding API to get location name
-            const response = await fetch(`https://nominatim.openstreetmap.org/reverse?format=json&lat=${latitude}&lon=${longitude}`)
-            const data = await response.json()
-
-            if (data.address) {
-              const city = data.address.city || data.address.town || data.address.village || ''
-              const country = data.address.country || ''
-              const locationString = city && country ? `${city}, ${country}` : city || country || 'Unknown Location'
-              setLocation(locationString)
-              setHasLocationBeenSet(true)
-            }
-          } catch {
-            setLocation('Location detected')
-            setHasLocationBeenSet(true)
-          } finally {
-            setIsLoadingLocation(false)
-            setShowLocationDropdown(false)
-          }
-        },
-        () => { setIsLoadingLocation(false) },
-        { enableHighAccuracy: true, timeout: 10000 }
-      )
-    } else {
-      setIsLoadingLocation(false)
-    }
-  }
-
-  // Request location on component mount
-  useEffect(() => {
-    // Don't set a default location immediately, let user choose or detect
-    // Automatically ask for location permission after a short delay
-    setTimeout(() => {
-      getUserLocation()
-    }, 1000)
-  }, [])
-
-  const handleLocationSelect = async (selectedLocation: string) => {
-    setLocation(selectedLocation)
-    setHasLocationBeenSet(true)
-    setShowLocationDropdown(false)
-    setCustomLocation('')
-    const geo = await geocodeLocation(selectedLocation)
-    if (geo) setCoords(geo)
-  }
-
-  const handleCustomLocationSubmit = async (e: React.FormEvent) => {
-    e.preventDefault()
-    if (customLocation.trim()) {
-      setLocation(customLocation)
-      setHasLocationBeenSet(true)
-      setShowLocationDropdown(false)
-      const geo = await geocodeLocation(customLocation.trim())
-      if (geo) setCoords(geo)
-      setCustomLocation('')
-    }
-  }
-
-  const geocodeLocation = async (loc: string): Promise<{ lat: number; lon: number } | null> => {
-    try {
-      const url = `https://nominatim.openstreetmap.org/search?format=json&q=${encodeURIComponent(loc)}`
-      const res = await fetch(url, { headers: { 'Accept-Language': 'en' } })
-      if (!res.ok) return null
-      const data = await res.json()
-      if (Array.isArray(data) && data.length > 0) {
-        const first = data[0]
-        return { lat: parseFloat(first.lat), lon: parseFloat(first.lon) }
-      }
-    } catch {}
-    return null
-  }
 
   const fetchHourlyWeather = async (lat: number, lon: number): Promise<string> => {
     try {
@@ -187,7 +106,7 @@ const HeroSection: React.FC<HeroSectionProps> = () => {
       const geo = await geocodeLocation(loc)
       if (geo) {
         activeCoords = geo
-        setCoords(geo)
+        // Note: coords are managed by useLocation hook, no need to set them here
       }
     }
 
@@ -359,83 +278,17 @@ RandomSeed: ${seed}`.trim()
           </h1>
 
           {/* Location selector */}
-          <div className="mb-8 inline-block relative z-40">
-            <div className="relative">
-              <button
-                onClick={() => setShowLocationDropdown(!showLocationDropdown)}
-                className="flex items-center gap-2 px-6 py-3 bg-white/20 backdrop-blur-md rounded-full text-white border border-white/30 hover:bg-white/30 transition-all duration-300"
-              >
-                <svg className="w-5 h-5" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                  <path
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    strokeWidth={2}
-                    d="M17.657 16.657L13.414 20.9a1.998 1.998 0 01-2.827 0l-4.244-4.243a8 8 0 1111.314 0z"
-                  />
-                  <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M15 11a3 3 0 11-6 0 3 3 0 016 0z" />
-                </svg>
-                <span className="font-medium">{isLoadingLocation ? 'Detecting location...' : location || 'Select location'}</span>
-                <svg
-                  className={`w-4 h-4 transition-transform ${showLocationDropdown ? 'rotate-180' : ''}`}
-                  fill="none"
-                  stroke="currentColor"
-                  viewBox="0 0 24 24"
-                >
-                  <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M19 9l-7 7-7-7" />
-                </svg>
-              </button>
-
-              {/* Dropdown menu */}
-              {showLocationDropdown && (
-                <div className="absolute z-50 top-full mt-2 left-1/2 -translate-x-1/2 w-64 bg-white/95 backdrop-blur-lg rounded-2xl shadow-2xl border border-white/50 overflow-visible">
-                  <div className="p-2">
-                    {/* FIXED button markup */}
-                    <button
-                      onClick={getUserLocation}
-                      className="w-full text-left px-4 py-3 hover:bg-gradient-to-r hover:from-orange-400/20 hover:to-purple-400/20 rounded-lg transition-colors flex items-center gap-2 text-gray-700"
-                      disabled={isLoadingLocation}
-                    >
-                      <svg className="w-4 h-4" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                        <path
-                          strokeLinecap="round"
-                          strokeLinejoin="round"
-                          strokeWidth={2}
-                          d="M15 10l4.553-2.276A1 1 0 0121 8.618v6.764a1 1 0 01-1.447.894L15 14M5 18h8a2 2 0 002-2V8a2 2 0 00-2-2H5a2 2 0 00-2 2v8a2 2 0 002 2z"
-                        />
-                      </svg>
-                      Use my current location
-                    </button>
-
-                    <div className="h-px bg-gradient-to-r from-transparent via-gray-300 to-transparent my-2" />
-
-                    <div className="max-h-48 overflow-y-auto">
-                      {popularLocations.map((loc) => (
-                        <button
-                          key={loc}
-                          onClick={() => handleLocationSelect(loc)}
-                          className="w-full text-left px-4 py-2 hover:bg-gradient-to-r hover:from-orange-400/20 hover:to-purple-400/20 rounded-lg transition-colors text-gray-700"
-                        >
-                          {loc}
-                        </button>
-                      ))}
-                    </div>
-
-                    <div className="h-px bg-gradient-to-r from-transparent via-gray-300 to-transparent my-2" />
-
-                    <form onSubmit={handleCustomLocationSubmit} className="p-2">
-                      <input
-                        type="text"
-                        placeholder="Enter custom location..."
-                        value={customLocation}
-                        onChange={(e) => setCustomLocation(e.target.value)}
-                        className="w-full px-3 py-2 bg-white/50 border border-gray-300 rounded-lg focus:outline-none focus:border-purple-400 text-gray-700"
-                      />
-                    </form>
-                  </div>
-                </div>
-              )}
-            </div>
-          </div>
+          <LocationSelector
+            location={location}
+            isLoadingLocation={isLoadingLocation}
+            showLocationDropdown={showLocationDropdown}
+            customLocation={customLocation}
+            setShowLocationDropdown={setShowLocationDropdown}
+            setCustomLocation={setCustomLocation}
+            getUserLocation={getUserLocation}
+            handleLocationSelect={handleLocationSelect}
+            handleCustomLocationSubmit={handleCustomLocationSubmit}
+          />
 
           {/* Open Map Button - Only show when location is set */}
           {hasLocationBeenSet && location && (
